Fixed performance problem in ResumeDownloader.allowAddition.  The problem was that IncompleteFileManager.getFile was called for every RemoteFileDesc received in response to a query.  As a result, IncompleteFileManager and downloads.dat could bloat with hashes for which no incomplete file existed.  Worse, allowAddition called IncompleteFileManager.getCompletedHash for every search result, which runs in linear time with respect to the number of hashes.
The solution is to add a same(..) method to IncompleteFileManager, which
simply compares names/sizes/hashes.  allowAddition uses this instead of
comparing the results of getFile; the two are exactly similar.  This change
prompted me to replace the _incompleteFileManager field of ResumeDownloader
with a simpler _hash field.  Added appropriate unit tests for same(..)
and serialization.