Fixed Uploader GUI problems and problems with UploaderTest
RouterService: fileManager and uploadManager are no longer final. UploadTest needs to
set these values. And ProvilegdedAccessor cannot change final fields.

UploadManager: HTTPUploader's constructor is called with the old HTTPUploader, this is so that the newly created HTTPUploader can get its BandwidthTrackerImpl from the older instance if possible so that the history is not lost.

HTTPUploader: Changed Constructor.

UploaderTest: Used the PrivilegedAccessor to set RouterService.fileManager to the FileManagerStub. and also at the beginning of each test set the uploadManager to a new instance of an UploadManager.

UploadDataLine: 
1. added a variable _cumulativeAmountUploaded, which remembers the number of bytes all the previous HTTPuploaders have uploaded.
2. set _size to the file size rather than the size of range requested.
3. When complete we dont set the progress to 100, because its possible to be asked for only a fraction of a file. However, there is a problem causing the progress to stop at 99. So if it is 99 we set it to 100. A little hacky but OK.
4. progress is sum of cumulative and currentUploader.amountRead / _size