UpdateManager requests update file and UploadManager gives it.
UpdateManager: Added code to find the ip and port of the node at other end - this can be a little tricky when its an incoming connection becuase the host is using an ephimeral port (we use OrigPort in Connection for this). Sends a HTTP request and reads file.

UploadManager: Changed so that these requests are treated like BrowseHost (ie no slot check and no queueing).

Added UpdateFileState : which does the upload for this kind of request.

Uploader: added int corresponding to the new state - UpdateFileState

HTTPUploader: added code to set the state for updateFile requests.