Removed all references to lib directory directly and moved update files to the user home dir.
ResourceManager: added a method to unpack the .ver file which contains the update.xml and the public.key file. The method unpacks the update.ver file into the user directory under 3 conditions
a) There is no update.xml file in the user dir
b) The version of limewire we are running is newer than the version in the update.xml file in the user directory. 
c) The update.xml in user directory says the version is @version@ 

Lazy Properties: If the UpdateManager thinks the latest version is @version@ then we dont want to 
send the X-Version header.

UpdateManager:
1. Picks up update.xml from the user home directory
2. isGreaterVersion signature has changed and its now public static
3. isGreaterVersion thinks @version@ is the greatest version.
4. If we download a new update.xml file it's done in the user home directory.

UpdateMessageVerifier: gets public.key from the user home directory.

UpdateFileState: gets update.xml from the user home directory