In establishConnection, we checked that we were not in some states before swithing to CONNECTING, but not the new HASHING and SAVING states. Now we check for these two states as well.
There was a race condition: the manager thread called start on the (worker) connectAndDownload thread, and then added it to threads. If the worker terminates before the it is added to threads, then it will be added to threads after it was removed (the remove operation is done by the worker - and in the race condition - would not mutate threads). We solve the racecondition by reversing the order- add first and start later. Thanks Sam!