Changes to core files to be compatible with LimePeerServer.
buildCoreJar.bat: added new packages to create core.jar properly

ConnectionManager: createConnectionBlocking would never get started because, a new OutgoinConnector was being created, but it was not being kicked off. This had no effect on the core code, because createConnectionBlocking was never called from the core. This seems wrong, since the method is supposed to block. However, I am not sure why it was creating a runnable if it was intended to be a blocking method. At any rate it gets the job done. 

StandardMessageRouter: changed sendResponses public, becasue it is public for ServerMessageRouter.

RouterService: Added a new constructor for the server to call.

LimeXMLDocument: re-added method setSchemaURI which was called by server.

LimeXMLUtils: re-added method creatResponses