DownloadTest: testStalledUploaderReplaced was not being being run, because we were setting the rate to 0, which was causing the "slow" TestUploader to crash before uploading anything -- we were not testing that slow uploaders are replaced.
ManagedDownloader: The idea of checking whether or not we are really http11 and using that in the same thread was not working out too well. The simple case of completing a download from one source was causing a spurious connection to be created. We now do things like we used to before, by caching the value of http11 at the start. If we find a http10 was actually a http11 rfd, the next time we use it, it will be used like a http11.