AlternateLocation, AlternateLocationsCollector: moved to altlocs package. AlternateLocationsCollection: 1.simplified so it does not check against a removed list 2. changed synchronization 3. made interface similar to a collection 4. moved to a new package and added todos.
AltLocCollectionsManager: Takes on some of the functionality that was earlier in AlternateLocationCollection (namely the removed list etc.). This class is not yet used by the other classes it is intended to be used by.

FileDesc, IncompleteFileDesc, HTTPDownloader, ManagedDownloader, HeadRequester, HTTPUploader: changed to access changed packages and method names in AltLoc related classes. Functionality changes follow...

Note: At this point we have taken away some functionality -- namely in the current state the code accepts altlocs without checking if they are bad etc...but that will change over the following commits