HTTPDownloader: 1. Removed completed TODOs. 2. removed unused method clearReceivedAltLocs.
AlternateLocationCollection: the httpStringValue method now writes out all the altlocs contained in it rather than just 10. This is how this will be used:

Downloaders: will send the entire collection and clear it out, this way the collection will always be current.

Uploaders(TODO): Will need to remember how many locations it sent, and create a new collection of size 10 from that point and write that out. This will roughly give new alternatelocations in successive requests.