MERGE FROM httpclient-branch
All active uses of HttpUrlConnection have been migrated over to HttpClient.
This fixes numerous problems because of bugs in Sun's implementation, such
as stalling if the remote-end did not send any response.  Even some of our
tests incorrectly relied on this behaviour, so it was a large problem.

There is a new project "httpclient-118" that will maintain our optimizations
and conversions within Jakarta Common's HttpClient code to keep it 1.1.8
compatable.

We also now use Common's Logging, which can wrap loggers in the order of:
 log4j, JDK1.4 logger, Common's Logging SimpleLog
 Default configuration files are included in this merge for jdk1.4 logger
 and SimpleLog.  Some classes have had Logs added to them to aid in debugging
 while fixing tests for the merge.

 Future debug information should use the Logger instead of multiple debug()
 calls.