This tests the following scenario: 1) LOCAL_IS_PRIVATE is true (as in shipped versions).  A connection from the local machine should not change the value of Acceptor.acceptedIncoming.
Currently this test fails because of the bug pointed out by G. Roper.  The
fix for this bug will soon follow.

This test could be expanded to do the following:
1) Test that a close IP does incoming does not change the value of
acceptedIncoming.
2) Currently, our incoming test does not try to read data from the incoming
connection.  This is probably not a huge deal, but the connect back spec
advises you check that you can actually read data from the incoming
connection.