Changes as per Sush's review comments. One more change coming...
ReplyHandler, Connection, ForMeReplyHandler, UPDReplyHandler, ReplyHandlerStub:
Changed the method name to send a StatisticVendorMessage in response to a GiveStatsVM. The method name is more consistent with the LimeWire way.

MessagesSupportedVendorMessage: 1. removed code to add StatisticVendorMessage is supported -- it's redundant given that GiveStatsVM is supported.
2. Added supportsGiveStatVM method which can be used before sending a GiveStatsVM down a tcp connection.

MessageRouter: De-threadified the handling of the GiveStatVM received.