ID3Editor: Changed to write the ID3v1 genre byte in the id3v2 frame only if it's a known/standard genre. Otherwise just write out the genre string.
ID3Reader:
Changed the code to assume that the closing brace of the id3v1 genre byte is either 2,3 or 4 of content of the genre id3v2 frame. If the brace is elsewhere or does not exist, we assume a non standard genre string and just read the whole thing.