Fixes the problem of us modifying the mp3 file that occured because the documents coming in from the network in the query replies which were cached and were calculated on the basis of id3v1.
Now, if there is ID3v2 data we will not overwrite it, but we will pick up fields from the query reply for which we have no value and add them to the currect document. The code also ensures that id3v2 that was being changed as a result of user editing, does not get subjected the same rules.