Fixing the issue caused by "half the id3 tags being written to disk, causing infinite looping".
The trick is to resolve the difference (if any) between the ID3data on disk at this stage, and the audio metadata in the filedesc. We use the id3tags from the disk and non id3tags from the audio metadata  in the file desc. 

ID3Reader: added utility method for finding if a tag is non-id3