Adding infrasctuture to allow Simpping for only certain settings.
LimeProps, AbstractSettings: made the getSettings method move up the inheritence hierarchy, so other types of settings can use it. 

Setting, (and it's subclasses): Added a final boolean for SimppEnabledness. Concrete Subclasses pass this boolean to super.

SettingsFactory: added method to create concrete subclasses of Setting with or without simpp-ness enabled. 

SimppSettingsManager: If we get a SimppSetting which is not simpp-enabled, skip it