Template for Mongolian (Cyrillic). Mongolian is an Altaic language spoken by approximately 5 millions people in Mongolia, China, Afghanistan and Russia, with related variants: Khalkha (or Halha), the national language of Mongolia, and Oirat, Chahar, and Ordos in the inner autonomous regions of China. Other languages in the same family include Buryat and Kalmyk (in Russia), and Moghul (or Mogul) in Afghanistan.
This file is for the Cyrillic-written Khalkha variant of Mongolia.

There also exists since 12th century (before it was banned in 1931), traditional Mongolian, using a vertical semitic alphabet with initial, medial and final letters, adpted from the Uighur alphabet, itself inherited from Sogdian, down from Aramaic. In 1931 the Latin alphabet was used temporarily, then in 1937 Mongolia was forced to use the Cyrillic alphabet under Russian pressure. Traditional Mongolian is used sometimes and morefreely in Mongolia today. But in China, the traditional Mongolian alphabet is used in autonomous regions, and never written with the Cyrillic alphabet. During the 13h to 15th century, Mongolian was also written with Han ideographs, and with Phags-Pa (derived from Tibetan).