Greatly simplify the notification to the gui:  * the GUI no longer polls on startup to see if an update was stored on disk.  Instead, while updates are being loaded we either schedule a notification or rely on a callback.  * If an update does not require a download, or we consider the download of the update hopeless, we schedule a notification containing the URL.  If the update does require a download and it was successful, we notify immediately.  * The gui distinguishes between updates that require visiting an URL by the presence of a command string.
We store a list of updates that failed to download between sessions and do not try to download an update that failed to download last time.  (That happens with a new setting StringSetSetting)  That list is updated whenever we get a new update message.

Simplify the calculation of the time the updates can start downloading.

Move a lot of things on the processing queue, making locking unnecessary and ensuring ordering.