Commit the new Indonesian update (number 2) sent by "N. Zulkarnaen" (n.zulkarnaen AT gmail DOT com). (53.23% complete)
After tests, needed to change the language code in Indonesian resource name from standard ISO 639-1 code "id" to legacy code "in", because Java silently replaces the language code in the Locale constructor, and causes the "MessagesBundle_id.properties" bundled not to be loaded with a Locale instanciated with the "id" code. This issue is similar to Hebrew code "iw" needed instead of "he".

Unfortunately, the locales resolver in Java's BundleProperties is still deficient to resolve such cases and still has no support for Locale aliases (such as those now being standardized in Unicode's CLDR project), and for languages that require distinction by script code (for those languages, one needs to use a variant code, but the resolution order is wrong)

Definitely, Java needs an extension of the Locale class, with a 4-parameters constructor, where the second one will be a script code. And implementation of Bundle requires enhancement to support Locale aliases and resolve them correctly.