Correct the typos signaled by Arne <Tux314@web.de> on October 16. Thanks.
But ignore ALL changes from FrostWire to LimeWire, and ignore the deletion of resources related to LimeWire Pro, the LimeWire team, and the LimeWire website.

Note that FrostWire is for now illegal, because it does not respect the Limewire copyright notice, and does not give credits to LimeWire, as REQUIRED by the GPL.

The GPL has no exception: the copyright and the GPL licence applies to all derivative works based on LimeWire original sources. And LimeWire can legally integrate any changes made in a derivative work based on Limewire work without asking for permission.