* revert to a single IncompleteDirectory Library tree node. * add some utility methods in IFM to check if a folder is a proper torrent folder * allow such folders to be deleted from the incomplete folder
Each incomplete torrent is stored in a subfolder of the incomplete folder.  The name of that
folder is the base32 encoded infohash of the torrent.  In that folder, there are only two entries:
the serialized infohash for easier resuming and the single file or folder for the torrent.

The serialized infohash has the same name as the torrent file/folder prefixed with .dat.  The reason
that is not stored in the IFM (like VerifyingFiles) is that infohashes can be really big and it would
be wasteful to keep them in memory.