BITT-22 Torrents are counted as an active upload, but not towards queued uploads.  Instead, they now have two additional states: SUSPENDED and AWAITING_REQUESTS.
The SUSPENDED state is entered when the torrent is complete, has someone interested in uploading from it but all uploads are choked.  The AWAITING_REQUESTS state is active when the torrent is active but not uploading because there is nobody interested in it, or it has no connections.

Also, the "Seeding" progress bar label now displays the upload:download ratio.