Code changes:
SnowMan: make consistent in handle method by adding null check for nodeId.
DartBoard: make consistent with other Painter subclasses by taking away public.

Javadocs:

Mojito-ui is pretty small so I included mojito db and db-imp for this commit and future CR review. For the db-imp classes, I took away "A default implementation ..." since the class name implied it - looked to answer "what is the class" and avoid the how of the class.

For a couple of classes, the class name perfectly described the class so I kept the documentation to a minimum (most manager classes).

Applied text and format consistency in documentation.