Made BTLinkManager use BTConnection in its interface instead of the BTLink interface.
This is necessary to provide access to the pieces each connection holds, this will allow us to use better strategies in piece selection.

Instead of selecting a piece at random, we can examine what pieces each connection has and determine which piece is rarest and download that if need be.