Finishing up refactoring of the swarmer code.
Broke out http dependencies in the swarmer.

Right now it uses the notion of registering a swarm source type with a handler. Which will know how to download from that source.

The only handler implemented right now is for downloading over http, in the future we could add a bittorrent handler, etc.