Initializer.preinit
Overriding look and feel for linux to be the corssplatform look and feel. mozswing will not work with GTKLookAndFeel on linux with a jvm less than 1.7

AppFrame.properties.
Telling jdesktop to use the look and feel already set in the UIManager. The lookAndFeel that the Initializer.preinit method sets should be used for jDesktop with this setting.