need to figure out guice with NavPanel, need to get TicTacToeFreindListPane to stop using ChatFriend and the new one. will need to simplify the dialog box that pops up, hopefully this will simplify all the copy and paste code so far.
this code has the start of the tic-tac-toe logic of playing. haven't added logic of a referee watching who's turn is next and then checking with the judge for a winner.

once the dialog box is created, need to get mouse clicks which then send the text to the referee who converts to x or o, checks for the winner, etc.