Re adds the capability to setManagedExtensions on the library back into the code.
Re adds the capability to set the Managed categories back into the code.

Adds the capability to remove a group of files from the various collections.

Moved logic to maintain fast accessible list of managed extensions for the managed categories to the LibraryFileData class. It will now rebuild the sets whenever a change is made.