Merging in changes from this review.https://www.limewire.org/fisheye/cru/LWCR-483
Libtorrent libraries built against 0.14.5

Changes to the native code to expose many additional libtorrent features.
Extensions to support getting additional peer data.
Extensions to support prioritizing individual files within the torrent.

Updates to support continuation of seeding torrents at application restart. Some of which was already on head.

Limits were put in, to the total number of seeding torrents, we are using libtorrents automanagement capabilities to ensure, the number of active uploading torrents never goes above a certain threshold.

The capability to select the seed to ratio has been exposed, this option is used to cancel seeding uploads when they reach this limit. This is to prevent torrents from staying in the lists forever.

The start and end listening ports have been exposed in the bittorrent options page.

The individual upload and download bandwidth sliders for torrents have been removed, instead torrents are using the applications sliders as a whole.