Merging bittorrent-5_3_1-branch.
https://www.limewire.org/fisheye/cru/LWCR-AP-95

The purpose of this patch is to support prioritizing files within the torrent to download as well as choosing not to download the file at all.

The user will be prompted to pick the files at the start of the torrent, or they may look at the file info for the torrent and pick the files there.