The main changes of note are getting rid of the ConnectionSpeed setting, and changing the max upload/download limits to be value based instead of percentage based. Everything else for the most part is just moving functionality around to different screens. As part of this fix, super really advanced no longer exists and has been moved to the advanced tab. The downloads tab, has become a "transfers" tab which gives more ready access to download/upload and bittorrent options.
This is a first pass, so while the ConnectionSpeed setting is no longer being used in the ui, it is still used in the back end. This can be changed, but is has some further reaching implications on how we do are tests, and how the app works as a whole. It would be better to see those changes in a separate review.