For the most part this is a refactoring of the torrent code.
Getting rid of the tight coupling between TorrentManager and the torrents themselves.
Cleaning up locking, making it more consistent.
Refactoring out limewire specific code, into a LimeWireTorrentManager that wraps the TorrentManagerImpl, which is ideally supposed to be a straight pass through to libtorrent for the most part.

Additionally TorrentDHTScheduler has been added along with supporting methods, to fail over to the mainline dht when a torrent cannot download. I am still cleaning up and evaluating the logic in this class. Also we technically we need to update some of the native code before this will work.