Updated dangerous file types: to protect older versions of LW from ASF files inside torrents, there are now no safe extensions for video/x-ms-asf. Future versions of LW will detect ASF files as video/x-lime-asf, which has the safe extensions previously used for video/x-ms-asf. A new mime type with no safe extensions, application/x-mal, was also added. Future versions of magic.mime can use this type to classify a file as malware.
One small but important change to the logic of FileExtensionChecker: if a file matches multiple dangerous mime types, it must have a safe extension for all of them in order to be considered safe. Previously a safe extension for the first type would have been sufficient. (No files are expected to match multiple dangerous mime types at present, but they may be in future with the addition of application/x-mal.)